STT/requirements.txt
# requirements
streamlit
openai
python-dotenv
langcodes